pyproject.toml
[project]
name = "mcp-aruba"
version = "0.2.0"
description = "MCP server for Aruba email and calendar via IMAP/SMTP/CalDAV"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    { name = "Giacomo Fiorucci", email = "giacomo.fiorucci@emotion-team.com" }
]

dependencies = [
    "mcp[cli]>=1.2.0",
    "python-dotenv>=1.0.0",
    "caldav>=1.3.9",
    "icalendar>=5.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.23.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mcp_aruba"]
